// backend/src/types.ts
export type Parameter = { parameter: string; value: any }